# [V1][TPU] Speed up top-k on TPU by using torch.topk