@hyeygit 2025-03-21
* Added env to guard the TPU-specific optimization
* Modified `test_sampler.py` to include top-k, but unable to verify the test as it crashes at head (filed #15257).