@hyeygit 2025-03-21
> Can you explain why there is a need to have an option to disable this?

Thanks for the review! This is a conservative measure -- in case there is any adverse effect from this change we can quickly disable it. This is meant to be temporary. Will send follow-up PR to remove once things look stable.