@NickLucche 2025-03-21
Thanks for the PR @hyeygit ! I did see that topk impl but I was (wrongly) afraid it would introduce dynamic shapes as the output has shape K, iirc.

As a general rule, I'd be in favor of testing PRs a bit more before merging rather than introducing temporary env variables to disable features.