@HAKSOAT 2025-03-22
Hi @hyeygit, thanks for your work here.

I was just studying your PR and I had a question.

If you the implementation says:

```
        if k is not None and p is None:
            # the implementation
        else:
            pass
```

Does this not mean that p would not have an effect on the sampler?

Or does it have an effect somehow that I am missing?