@njhill 2025-03-25
@hyeygit I'm curious whether this was tested at all? It looks completely wrong to me - torch.topk only takes a scalar and k here is a batch-size vector, which also may have entries equal to the vocab size (i.e. top k disabled). Also as @HAKSOAT points out in the current state top-p is ignored.

You can still use torch.topk with the max top k from the batch and gather the values from the result, and this is much faster than doing the whole sort - for GPU too. I have opened a PR for this https://github.com/vllm-project/vllm/pull/15478, perhaps the TPU-specific path is no longer needed after that (until there's a dedicated kernel)?